+++ src/caselawclient/models/identifiers/exceptions.py
@@ -2,9 +2,5 @@ class InvalidIdentifierXMLRepresentationException(Exception):
     pass
 
 
-class UUIDMismatchError(Exception):
-    pass
-
-
 class IdentifierValidationException(Exception):
     pass